+++ CHANGELOG.md
@@ -3,9 +3,17 @@
 Notable changes to  Format based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). Project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 <!--
+-->
+
 ## [Unreleased]
 
--->
+### Added
+
+- demo `demo/zipcode.py`
+
+### Fixed
+
+- async tool definitions
 
 ## [0.4.0] - 20240802